Split the two e2e tests that set the pipeline's floor
The e2e phase is gated entirely by its two longest files: quiz_auth at 11.6
minutes and sitemap at 11.2, while every other shard finishes by 11:15. Packing
the rest more tightly cannot help until these two come down.

sitemap splits cleanly three ways. Each shard samples independently and visits
its share, so the number of pages visited per run is unchanged.

quiz_auth only splits two ways, along the fixture seam it already had. #1309
merged this file together to conserve TOTP codes, and that constraint is real
but narrower than it looks: the file spends 3 codes and 33 seconds waiting out
of 11.6 minutes, so the rest is genuine work that parallelizes. What stops us
going further is that the code service issues one code per period, so shards
serialize on it -- N shards cost roughly 1.5N minutes of waiting while dividing
the work by N, and past two that stops paying off. The eslint rule stays as the
place that reasoning is written down.

Neither file takes screenshots, so no reference images move.
