+++ react/test/sitemap_x0.test.ts
@@ -1,0 +1,3 @@
+import { visitSitemapLinks } from './sitemap_test_template'
+
+visitSitemapLinks(0, 3)
