+++ react/test/sitemap_x1.test.ts
@@ -1,0 +1,3 @@
+import { visitSitemapLinks } from './sitemap_test_template'
+
+visitSitemapLinks(1, 3)
