+++ react/test/sitemap_x2.test.ts
@@ -1,0 +1,3 @@
+import { visitSitemapLinks } from './sitemap_test_template'
+
+visitSitemapLinks(2, 3)
